Write the case management deployments

Write the case management deployments

Usage

defineChangeHS(
  baseList,
  name = "Change in case management",
  startDate = NULL,
  endDate = NULL,
  interval = NULL,
  dates = NULL,
  initACT = 1,
  initQN = 1,
  initSelf = 1,
  compACT = 1,
  compQN = 1,
  compSelf = 1,
  pSeekOfficialCareUncomplicated1,
  pSelfTreatUncomplicated = 0.01821375,
  pSeekOfficialCareUncomplicated2,
  pSeekOfficialCareSevere = 0.48
)

define_changeHS(
  baseList,
  name = "Change in case management",
  startDate = NULL,
  endDate = NULL,
  interval = NULL,
  dates = NULL,
  initACT = 1,
  initQN = 1,
  initSelf = 1,
  compACT = 1,
  compQN = 1,
  compSelf = 1,
  pSeekOfficialCareUncomplicated1,
  pSelfTreatUncomplicated = 0.01821375,
  pSeekOfficialCareUncomplicated2,
  pSeekOfficialCareSevere = 0.48
)

Arguments

baseList

List with experiment data.

name

Name of the intervention

startDate

Date in YYYY-MM-DD format.

endDate

Date in YYYY-MM-DD format.

interval

A string like '1 weeks'. Same as in seq.Date(). Or a list composed of the entries 'days' (optional), 'months' (optional) and 'years'. If a list is used, startDate and endDate are not used and can be NULL.

dates

If NULL, startDate, endDate and interval are used, else a vector of dates in YYYY-MM-DD format. Can be a placeholder.

initACT

Initial artemisinine combination therapy. Can be a placeholder.

initQN

Initial quinine. Can be a placeholder.

initSelf

Initial probability of self-treatment. Can be a placeholder.

compACT

Compliance artemisinine combination therapy. Can be a placeholder.

compQN

Compliance quinine. Can be a placeholder.

compSelf

Compliance to self-treatment. Can be a placeholder.

pSeekOfficialCareUncomplicated1

Probability that a patient with newly incident uncomplicated disease seeks official care. Can be a placeholder.

pSelfTreatUncomplicated

Probability that a patient with uncomplicated disease without recent history of disease (i.e. first line) will self-treat. Can be a placeholder.

pSeekOfficialCareUncomplicated2

Probability that a patient with recurrence of uncomplicated disease seeks official care. Can be a placeholder.

pSeekOfficialCareSevere

Probability that a patient with severe disease obtains appropriate care. Can be a placeholder.