Writes the health system part

Writes the health system part

Usage

defineHealthSystem(
  baseList,
  name = "Tanzania ACT",
  initACT = 1,
  initQN = 1,
  initSelf = 1,
  compACT = 1,
  compQN = 1,
  compSelf = 1,
  pSeekOfficialCareUncomplicated1,
  pSelfTreatUncomplicated = 0.01821375,
  pSeekOfficialCareUncomplicated2,
  pSeekOfficialCareSevere = 0.48
)

define_health_system(
  baseList,
  name = "Tanzania ACT",
  initACT = 1,
  initQN = 1,
  initSelf = 1,
  compACT = 1,
  compQN = 1,
  compSelf = 1,
  pSeekOfficialCareUncomplicated1,
  pSelfTreatUncomplicated = 0.01821375,
  pSeekOfficialCareUncomplicated2,
  pSeekOfficialCareSevere = 0.48
)

Arguments

baseList

List with experiment data.

name

Name of health system

initACT

Initial artemisinine combination therapy.

initQN

Initial quinine.

initSelf

Initial probability of self-treatment.

compACT

Compliance artemisinine combination therapy.

compQN

Compliance quinine.

compSelf

Compliance to self-treatment.

pSeekOfficialCareUncomplicated1

Probability that a patient with newly incident uncomplicated disease seeks official care.

pSelfTreatUncomplicated

Probability that a patient with uncomplicated disease without recent history of disease (i.e. first line) will self-treat.

pSeekOfficialCareUncomplicated2

Probability that a patient with recurrence of uncomplicated disease seeks official care.

pSeekOfficialCareSevere

Probability that a patient with severe disease obtains appropriate care.