Function that writes the larviciding component

Function that writes the larviciding component

Usage

defineLarv(
  baseList,
  mosquitos,
  component = "LSM",
  coverage = "@futLSMcov@",
  decayVals = list(L = 0.25, k = NULL, funct = "step"),
  startDate = NULL,
  endDate = NULL,
  interval = NULL,
  dates = NULL
)

define_larv(
  baseList,
  mosquitos,
  component = "LSM",
  coverage = "@futLSMcov@",
  decayVals = list(L = 0.25, k = NULL, funct = "step"),
  startDate = NULL,
  endDate = NULL,
  interval = NULL,
  dates = NULL
)

Arguments

baseList

List with experiment data.

mosquitos

Names of mosquito species to simulate.

component

Name of vector population intervention (i.e. LSM).

coverage

A variable or a numeric value. Can be a placeholder.

decayVals

Decay (if not specified, constant effectiveness, step function as default)

startDate

Date in YYYY-MM-DD format.

endDate

Date in YYYY-MM-DD format.

interval

A string like '1 weeks'. Same as in seq.Date(). Or a list composed of the entries 'days' (optional), 'months' (optional) and 'years'. If a list is used, startDate and endDate are not used and can be NULL.

dates

If NULL, startDate, endDate and interval are used, else a vector of dates in YYYY-MM-DD format. Can be a placeholder.