Adds vaccine intervention parameterisation to baseList

Adds vaccine intervention parameterisation to baseList

Usage

defineVaccine(
  baseList,
  vaccineParameterization,
  append = TRUE,
  name = NULL,
  verbatim = FALSE,
  hist = FALSE
)

define_vaccine(
  baseList,
  vaccineParameterization,
  append = TRUE,
  name = NULL,
  verbatim = FALSE,
  hist = FALSE
)

Arguments

baseList

List with experiment data.

vaccineParameterization

Parameterization list

append

If TRUE, then append to existing baseList, otherwise overwrites

name

Name tag list

verbatim

If TRUE, then show messages

hist

If TRUE, then decay is assumed to be step function set to 1 for a year and then to zero for the remainder