Deployment of an continuous intervention

Deployment of an continuous intervention

Usage

deploy_cont_compat(
  baseList,
  component = "IPTi",
  begin = "2019-01-01",
  end = "2030-01-01",
  targetAgeYrs = NULL,
  vaccMinPrevDoses = NULL,
  vaccMaxCumDoses = NULL,
  coverage = NULL,
  varyCov = FALSE,
  restrictToSubPop = NULL
)

Arguments

baseList

List with experiment data.

component

Name of intervention

begin

Start date of the intervention (yyyy-mm-dd)

end

End date of the intervention (yyyy-mm-dd)

targetAgeYrs

Age at which intervention is received (1=1 year) e,g. c(0.25,0.33 ,0.75)

vaccMinPrevDoses

Applies to vaccines only: vaccine doses are only deployed by this deployment if the previous number of doses (for the component deployed) is at least this number. Needs to be of the same length as targetAgeYrs.

vaccMaxCumDoses

Applies to vaccines only: vaccine doses are only deployed by this deployment if the previous number of doses (for the component deployed) is less than this number. Needs to be of the same length as targetAgeYrs.

coverage

Coverage, either fixed (integer) or as a variable ("@IPTcov@"), same order as for target age years

varyCov

Default is FALSE , for varying coverage

restrictToSubPop

f this element is specified, deployment is restricted to some sub-population