Run openMalaria scenarios

Generates a system call to launch OpenMalaria with the correct parameters for each scenario found in the directory. It is assumed that OpenMalaria is available in the PATH.

Usage

runSimulations(
  scenarios = NULL,
  cmd = "openMalaria",
  dryRun = FALSE,
  verbose = FALSE,
  ncores = 1,
  rowStart = NULL,
  rowEnd = NULL
)

run_simulations(
  scenarios = NULL,
  cmd = "openMalaria",
  dryRun = FALSE,
  verbose = FALSE,
  ncores = 1,
  rowStart = NULL,
  rowEnd = NULL
)

Arguments

scenarios

Scenario data frame

cmd

Command to run openMalaria.

dryRun

If TRUE, only the final command is printed but not executed.

verbose

If TRUE, use OpenMalaria's verbose output.

ncores

Number of parallel processes to use.

rowStart

Starting row. Optional.

rowEnd

End row. Optional.