Function to write the case management deployments

Function to write the case management deployments

Usage

define_changeHS_compat(
  baseList,
  access = "Access",
  coverage = NULL,
  y1 = 2000,
  y2 = 2015,
  use_at_symbol = TRUE,
  pSelfTreatUncomplicated = 0.01821375,
  futSevere = NULL,
  pSeekOfficialCareSevere = 0.48,
  SIMSTART = "1918-01-01",
  every = 1,
  interval = "year",
  m1 = 1,
  m2 = 1,
  init.act = 1,
  init.qn = 1,
  init.self = 1,
  comp.act = 1,
  comp.qn = 1,
  comp.self = 1
)

Arguments

baseList

List with experiment data.

access

Name of coverage value (i.e. "Access")

coverage

Name of future intervention for change in HS (i.e. "@futCM@")

y1

Year of the first date (surveys starting from year y1)

y2

Year of the end date (surveys continuing until year y2)

use_at_symbol

If TRUE, variable will be replaced with setting specific values

pSelfTreatUncomplicated

Proportion of uncomplicated episodes treated at home

futSevere

Variable for different values of 'pseekOfficialCareSevere'

pSeekOfficialCareSevere

Proportion of severe episodes that seek care at the formal sector

SIMSTART

Start of the simulations (equal to ORIGIN!)

every

Interval size

interval

Interval size (days, weeks, )

m1

Month of the first date

m2

Month of the end date

init.act

Initial ACT

init.qn

Initial Quinine

init.self

Initial self treatment

comp.act

See OpenMalaria

comp.qn

See OpenMalaria

comp.self

See OpenMalaria